Streamline table styles, use better defaults
Tables are the most important part of ActiveAdmin as it’s the
components that is most widely used to display data. Keep presentation
clean and consistent, same padding, colors, etc. We can fine tune but
this brings it down just to the necessary styles without being verbose.
Further improvements: removing even/odd classes as this can now be done
using solely CSS.

Height acts as min-height on a table or table cell. Forgot about this.
I switched to min-height since its so much clearer but this is how it
works so lets keep it for now.

Change table stripe and selected colors. A soft yellow is a better
choice I think as a default for the selected background color. I don’t
think $primary-color is a good choice for striped backgrounds as a
default.

Adjust table styles to account for various contexts. Panel tables
inherit styles even for the attributes_table. I’d rather avoid
resetting a lot of styles so going to consider adding a new class
if possible. For now this is sufficient.

Set table columns to vertical align top. The table header columns
though are set to align bottom.