Streamline pagination styles
Remove styles that have no effect. Make a few tweaks so all content
lines up horizontally. Don’t customize select dropdown as we don’t do
that elsewhere so keep it consistent (we could see about changing this
to our dropdown component but this for sure affects Kaminari and we
already have an issue with their templates).