Bring back tabs styles and streamline
I removed this because it was poorly named so when I searched for
ui-tab-nav nothing came up. That class name is used by jQuery UI but we
should be using our own (which is tabs.rb sets as “nav-tabs”) to style
the component otherwise its confusing if its being used or not.