Add explicit dialog buttons so we can add classes
In the long run we’d like to style all buttons via a reusable set of
classes rather than have to target each element uniquely and use a
mixin. Buttons are so common it would be great to have a reusable class
especially if someone is customizing ActiveAdmin with custom
interactions and needs buttons.