Remove excessive status_tag classes
Confusing to have so many, keep it to just a select few so its
consistent. Removed: warn, orange, errored, red, complete, published,
and green. For the ok status I left in ok and completed (this probably
should be “success”).

I think the default status_tag background color should be gray, the
same as the “no” class but we can decide that later once configuring
all values.