New sass variables for base, flash msg, and buttons
This is a good first start. Started with setting the global base
variables we’ll use across the app. Using Bootstrap/Foundation as
inspiration for naming and what to configure. With this change I
increased the base font size to 13px which feels a lot better now.
Button variables only are for size, haven’t done colors yet.