Update blank_slate component to accept child elements
Now works pretty much the same as a Panel component. This way we can
use `blank_slate` with a block and specify whatever elements we like
rather than passing a content string (although that is still
supported). Now our dashboard template is much cleaner as we use the
built in component rather than hard code CSS and styles.