Update default theme with Flat UI and cleanup variables
The previous color scheme didn’t feel very in sync so I’ve used the
Flat UI color theme here across the board. I’ve cleaned up variables so
we only have now the new scheme going forward. Variables that were
removed or renamed are now in a commented section at the top.

I’ve configured most of the Flat UI theme colors we use
in variables with the same matching name from the theme. Some are only
used once but figured it’d be good to just configure them by name so
its easier to find which ones we are using if someone else is curious
or wants to contribute.