Replace #login with better named class and simplify styles
No need to override <ol> element as we now have consistent padding on
fieldset. No need for link style since we include the secondary
authentication link as a secondary action in the form (sticks with
formtastic’s style by default). This should be good enough for a
default and then others can customize as they see fit.

Replace #login with a .devise-login-container class.