Use correct class name for dropdown component
This was a name change I made since it was confusing to me. The menu
portion is what pops out. The list of options. The dropdown itself is
the whole component. I should have been more careful on rebasing and
spotting this but in my defense its been a year now since I’ve worked
on this feature branch. Since then I’ve done occasional rebases.