Fix create_another support
No need to specify a custom block for the form actions in the form
component. I thought I did at first to get the default rendering I
wanted but this seems to be fine due to later changes. I learned a lot
about ActiveAdmin as I spent time working on this as I was unfamiliar
with its setup.

Due to our drastic CSS changes most of the create_another styles are
not necessary since we’ve simplified and removed unused styles.