Use `helpers.tag` to prevent `tag` clash with Tag model
No idea how these are clashing since one is lowercase but there is some
crazy dynamic stuff going on here. From what I can tell there is no way
to access Rails helpers via a unique scope, its just referenced
globally. The issue I ran into is on the generated test app, we have a
Tag model and an admin page for it. When viewing the /tags/new page it
raises an error on this line but it doesn’t do it for another model at
the new page. It has to be a name clash. Further I have no idea how to
get arbre to work nicely with Rails helpers and it can be a massive
PITA. Best and simplest way I could resolve this issue.