Inherit font styles on form controls
Normalize went through some changes on v6 and v7 and since v7 is the
latest it sets a default font family and line height but we don’t want
that. While it is normalizing styles, we’d prefer that form controls
inherit so they pick up the custom font we set as a default or a user
sets on their end.